--- src/test/java/school/redrover/TestV.java
@@ -1,30 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-public class TestV {
-    @Test
-    public void testTest2() {
-        WebDriver driver = new ChromeDriver();
-        driver.get("https://www.saucedemo.com/");
-
-        WebElement username = driver.findElement(By.id("user-name"));
-        username.sendKeys("standard_user");
-
-        WebElement password = driver.findElement(By.id("password"));
-        password.sendKeys("secret_sauce");
-
-        WebElement submitButton = driver.findElement(By.id("login-button"));
-        submitButton.click();
-
-        String pageTitle = driver.getTitle();
-        Assert.assertEquals(pageTitle, "Swag Labs");
-
-        driver.quit();
-    }
-}
